Made debuggign easier, changed to tabs, improved readabillity, started functions
Imported the logging module which was used to replace the log()
function.

Improved the readabillity of the file slightly.

The program should call a function when the user gives it information.
So if they have a burn, it should call the function burn(). This will
make changes easier in the future nad will also keep the code clean. I'm
not sure whether the code should stop after the user has seen the burn()
message or not. Up for discussion.