Put function inside decorator
I think the functions should beinside the decorator as the decorator
("the @ symbol") is called and does the main brunt force of the
messaging app. By putting all functions inside the decorator and
treating the decorator as a mini-program I suspect that the function
burn() will work. I will test this now.